+++ src/app/dashboard/almacenes/[id]/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useEffect, useState } from "react";
-import { jsonOrNull } from "@lib/http";
 import { useParams } from "next/navigation";
+import { jsonOrNull } from "@lib/http";
 import Spinner from "@/components/Spinner";
 import MaterialRow, { Material } from "../components/MaterialRow";
 
@@ -18,12 +18,13 @@ interface Almacen {
   inventario?: number;
 }
 
-
-export default function AlmacenDetallePage() {
-  const params = useParams();
-  const id = params.id as string;
+export default function AlmacenPage() {
+  const { id } = useParams();
   const [almacen, setAlmacen] = useState<Almacen | null>(null);
-  const [filas, setFilas] = useState<Material[]>([]);
+  const [materiales, setMateriales] = useState<Material[]>([]);
+  const [seleccion, setSeleccion] = useState<number | null>(0);
+  const [busqueda, setBusqueda] = useState("");
+  const [orden, setOrden] = useState<"producto" | "cantidad">("producto");
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState("");
 
@@ -34,31 +35,19 @@ export default function AlmacenDetallePage() {
       .then((data) => {
         if (data.error) throw new Error(data.error);
         setAlmacen(data.almacen);
-        setFilas(data.almacen?.inventarioDetalle || []);
+        setMateriales(data.almacen?.inventarioDetalle || []);
+        if ((data.almacen?.inventarioDetalle || []).length === 0) {
+          setSeleccion(null);
+        }
       })
       .catch(() => setError("Error al cargar almacén"))
       .finally(() => setLoading(false));
   }, [id]);
 
-  if (error)
-    return (
-      <div className="p-4 text-red-500" data-oid="mmox.wc">
-        {error}
-      </div>
-    );
-
-  if (loading)
-    return (
-      <div className="p-4" data-oid="60abg5i">
-        <Spinner />
-      </div>
-    );
-
-  if (!almacen)
-    return (
-      <div className="p-4" data-oid="hf-4vbh">
-        No encontrado
-      </div>
+  const filtrados = materiales
+    .filter((m) => m.producto.toLowerCase().includes(busqueda.toLowerCase()))
+    .sort((a, b) =>
+      orden === "producto" ? a.producto.localeCompare(b.producto) : a.cantidad - b.cantidad,
     );
 
   const actualizar = (
@@ -66,27 +55,49 @@ export default function AlmacenDetallePage() {
     campo: keyof Material,
     valor: string | number,
   ) => {
-    setFilas((f) => {
-      const arr = [...f];
+    setMateriales((ms) => {
+      const arr = [...ms];
       // @ts-ignore
       arr[idx][campo] = campo === "cantidad" ? Number(valor) : valor;
       return arr;
     });
   };
 
-  const agregarFila = () => {
-    setFilas((f) => [...f, { producto: "", cantidad: 0, lote: "" }]);
+  const guardar = () => {
+    alert("Guardado");
+  };
+  const cancelar = () => setSeleccion(null);
+  const duplicar = () => {
+    if (seleccion === null) return;
+    setMateriales((ms) => [...ms, { ...ms[seleccion] }]);
   };
 
+  if (loading)
+    return (
+      <div className="p-4">
+        <Spinner />
+      </div>
+    );
+
+  if (error)
+    return (
+      <div className="p-4 text-red-500">
+        {error}
+      </div>
+    );
+
+  if (!almacen)
+    return (
+      <div className="p-4">
+        No encontrado
+      </div>
+    );
+
   return (
-    <div className="space-y-4" data-oid="0hw.ll8">
-      <h1 className="text-2xl font-bold" data-oid="1zoe091">
-        {almacen.nombre}
-      </h1>
+    <div className="space-y-4">
+      <h1 className="text-2xl font-bold">{almacen.nombre}</h1>
       {almacen.descripcion && (
-        <p className="text-sm text-[var(--dashboard-muted)]" data-oid="1ainm7s">
-          {almacen.descripcion}
-        </p>
+        <p className="text-sm text-[var(--dashboard-muted)]">{almacen.descripcion}</p>
       )}
       {almacen.encargado && (
         <p className="text-xs text-[var(--dashboard-muted)]">
@@ -96,8 +107,7 @@ export default function AlmacenDetallePage() {
       )}
       {almacen.ultimaActualizacion && (
         <p className="text-xs text-[var(--dashboard-muted)]">
-          Última actualización:{' '}
-          {new Date(almacen.ultimaActualizacion).toLocaleString()}
+          Última actualización: {new Date(almacen.ultimaActualizacion).toLocaleString()}
         </p>
       )}
       {typeof almacen.inventario === 'number' && (
@@ -106,27 +116,102 @@ export default function AlmacenDetallePage() {
         </p>
       )}
 
-      <table className="w-full text-sm bg-white/5 rounded-md overflow-hidden">
-        <thead className="bg-white/10">
-          <tr>
-            <th className="px-3 py-2 text-left">Producto</th>
-            <th className="px-3 py-2 text-left">Cantidad</th>
-            <th className="px-3 py-2 text-left">Lote</th>
-          </tr>
-        </thead>
-        <tbody>
-          {filas.map((f, idx) => (
-            <MaterialRow key={idx} material={f} index={idx} onChange={actualizar} />
-          ))}
-        </tbody>
-      </table>
-
-      <button
-        onClick={agregarFila}
-        className="px-4 py-2 bg-[var(--dashboard-accent)] text-white rounded-lg hover:bg-[var(--dashboard-accent-hover)] text-sm"
-      >
-        Agregar fila
-      </button>
+      <div className="flex h-full" data-oid="inv.layout">
+        <aside className="w-72 max-w-sm border-r border-white/10 p-4 space-y-4">
+          <div className="flex gap-2">
+            <input
+              value={busqueda}
+              onChange={(e) => setBusqueda(e.target.value)}
+              placeholder="Buscar"
+              className="flex-1 p-2 rounded-md bg-white/5 focus:outline-none"
+            />
+            <select
+              value={orden}
+              onChange={(e) => setOrden(e.target.value as any)}
+              className="p-2 rounded-md bg-white/5"
+            >
+              <option value="producto">Nombre</option>
+              <option value="cantidad">Cantidad</option>
+            </select>
+          </div>
+          <ul className="space-y-1 overflow-y-auto max-h-[calc(100vh-12rem)]">
+            {filtrados.map((m, idx) => (
+              <li key={idx}>
+                <button
+                  onClick={() => setSeleccion(idx)}
+                  className={`w-full text-left p-2 rounded-md transition ${idx === seleccion ? 'bg-white/10' : 'hover:bg-white/5'}`}
+                >
+                  {m.producto}
+                </button>
+              </li>
+            ))}
+          </ul>
+          <div className="flex gap-2">
+            <button
+              onClick={() =>
+                setMateriales((ms) => [...ms, { producto: '', cantidad: 0, lote: '' }])
+              }
+              className="flex-1 py-1 rounded-md bg-[var(--dashboard-accent)] text-white text-sm hover:bg-[var(--dashboard-accent-hover)]"
+            >
+              Nuevo
+            </button>
+            <button
+              onClick={duplicar}
+              disabled={seleccion === null}
+              className="flex-1 py-1 rounded-md bg-white/10 text-white text-sm disabled:opacity-50"
+            >
+              Duplicar
+            </button>
+          </div>
+        </aside>
+        <section className="flex-1 p-4 space-y-4 overflow-y-auto">
+          {seleccion === null ? (
+            <p className="text-sm text-[var(--dashboard-muted)]">
+              Selecciona un material para editar.
+            </p>
+          ) : (
+            <>
+              <table className="w-full text-sm bg-white/5 rounded-md overflow-hidden">
+                <thead className="bg-white/10">
+                  <tr>
+                    <th className="px-3 py-2 text-left">Producto</th>
+                    <th className="px-3 py-2 text-left">Cantidad</th>
+                    <th className="px-3 py-2 text-left">Lote</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  <MaterialRow
+                    material={materiales[seleccion]}
+                    index={seleccion}
+                    onChange={actualizar}
+                  />
+                </tbody>
+              </table>
+              <div className="flex gap-2">
+                <button
+                  onClick={guardar}
+                  className="px-4 py-2 rounded-lg bg-[var(--dashboard-accent)] text-white text-sm hover:bg-[var(--dashboard-accent-hover)]"
+                >
+                  Guardar
+                </button>
+                <button
+                  onClick={cancelar}
+                  className="px-4 py-2 rounded-lg bg-white/10 text-white text-sm"
+                >
+                  Cancelar
+                </button>
+                <button
+                  onClick={duplicar}
+                  className="px-4 py-2 rounded-lg bg-white/10 text-white text-sm"
+                >
+                  Duplicar
+                </button>
+              </div>
+            </>
+          )}
+        </section>
+      </div>
     </div>
   );
 }
+
